Add parcelization to screen objects
