+++ app-launcher/android/src/main/kotlin/io/ashdavies/playground/LauncherActivity.kt
@@ -1,13 +1,8 @@
 package io.ashdavies.playground
 
-import android.os.Parcelable
 import androidx.activity.compose.setContent
 import androidx.core.view.WindowCompat.setDecorFitsSystemWindows
 import com.arkivanov.decompose.defaultComponentContext
-import kotlinx.parcelize.Parcelize
-
-@Parcelize
-internal object AndroidLauncherScreen : LauncherScreen, Parcelable
 
 internal class LauncherActivity : KotlinActivity({
     setDecorFitsSystemWindows(window, true)
@@ -15,7 +10,7 @@ internal class LauncherActivity : KotlinActivity({
     val circuitConfig = CircuitConfig(defaultComponentContext())
     val initialBackStack = buildInitialBackStack(
         nextScreen = intent.getStringExtra("route"),
-        initialScreen = AndroidLauncherScreen,
+        initialScreen = LauncherScreen,
     )
 
     setContent {
@@ -24,4 +19,4 @@ internal class LauncherActivity : KotlinActivity({
             circuitConfig = circuitConfig,
         )
     }
-},)
+})
