+++ dominion-app/src/commonMain/kotlin/io/ashdavies/dominion/DominionPresenter.kt
@@ -1,5 +1,11 @@
 package io.ashdavies.dominion
 
+import com.arkivanov.essenty.parcelable.Parcelable
+import com.arkivanov.essenty.parcelable.Parcelize
 import com.slack.circuit.Screen
 
-public object DominionScreen : Screen
+// import io.ashdavies.parcelable.Parcelable
+// import io.ashdavies.parcelable.Parcelize
+
+@Parcelize
+public object DominionScreen : Parcelable, Screen
