+++ events-app/src/commonMain/kotlin/io/ashdavies/playground/EventsPresenter.kt
@@ -1,5 +1,11 @@
 package io.ashdavies.playground
 
+import com.arkivanov.essenty.parcelable.Parcelable
+import com.arkivanov.essenty.parcelable.Parcelize
 import com.slack.circuit.Screen
 
-public object EventsScreen : Screen
+// import io.ashdavies.parcelable.Parcelable
+// import io.ashdavies.parcelable.Parcelize
+
+@Parcelize
+public object EventsScreen : Parcelable, Screen
